+++ app/frontend/src/components/Buttons/_index.scss
@@ -1,2 +1,2 @@
-@forward "Buttons";
+@forward "Button";
 @forward "CloseButton";
