+++ app/frontend/src/components/Navigation/HeaderNav.js
@@ -5,7 +5,7 @@ import React, { Fragment, useState } from "react";
 // Internal Imports
 import * as utility from "../Utility/utils";
 import { IconHamburgerMenu } from "../../assets/images/images";
-import { Button } from "../Buttons/Buttons";
+import { Button } from "../Buttons/Button";
 
 function HeaderNav({ logoDesktop, logoMobile, menu }) {
   const [open, setOpen] = useState(false);
